Fixed bug where prefetching-cache was not purged when a new page was created and was automatically added to a menu
