+++ src/Servebolt/ServeboltOptimizer.php
@@ -4,34 +4,33 @@ namespace Servebolt\Optimizer;
 
 if (!defined('ABSPATH')) exit; // Exit if accessed directly
 
-use Servebolt\Optimizer\Admin\ClearSiteDataHeader\ClearSiteDataHeader;
-use Servebolt\Optimizer\CachePurge\WpCachePurge;
-use Servebolt\Optimizer\Prefetching\WpPrefetching;
-use Servebolt\Optimizer\Compatibility\Compatibility as PluginCompatibility;
 use Servebolt\Optimizer\AcceleratedDomains\AcceleratedDomains;
+use Servebolt\Optimizer\AcceleratedDomains\Prefetching\WpPrefetching;
+use Servebolt\Optimizer\Admin\AdminBarGui\AdminBarGui;
+use Servebolt\Optimizer\Admin\AdminController;
+use Servebolt\Optimizer\Admin\Assets as AdminAssets;
+use Servebolt\Optimizer\Admin\ClearSiteDataHeader\ClearSiteDataHeader;
+use Servebolt\Optimizer\AssetAutoVersion\AssetAutoVersion;
+use Servebolt\Optimizer\CachePurge\WpCachePurge;
+use Servebolt\Optimizer\Cli\Cli;
+use Servebolt\Optimizer\CloudflareImageResize\CloudflareImageResize;
+use Servebolt\Optimizer\Compatibility\Compatibility as PluginCompatibility;
 use Servebolt\Optimizer\FullPageCache\FullPageCache;
-use Servebolt\Optimizer\MenuOptimizer\WpMenuOptimizer;
 use Servebolt\Optimizer\GenericOptimizations\GenericOptimizations;
+use Servebolt\Optimizer\MenuOptimizer\WpMenuOptimizer;
+use Servebolt\Optimizer\PluginActiveStateHandling\PluginActiveStateHandling;
+use Servebolt\Optimizer\Queue\QueueParseEventHandler;
 use Servebolt\Optimizer\TextDomainLoader\WpTextDomainLoader;
 use Servebolt\Optimizer\Utils\Crypto\OptionEncryption;
-use Servebolt\Optimizer\CloudflareImageResize\CloudflareImageResize;
-use Servebolt\Optimizer\Queue\QueueParseEventHandler;
 use Servebolt\Optimizer\WpCron\WpCronCustomSchedules;
 use Servebolt\Optimizer\WpCron\WpCronEvents;
-use Servebolt\Optimizer\Admin\AdminBarGui\AdminBarGui;
-use Servebolt\Optimizer\Admin\Assets as AdminAssets;
-use Servebolt\Optimizer\Admin\AdminController;
-use Servebolt\Optimizer\AssetAutoVersion\AssetAutoVersion;
-use Servebolt\Optimizer\Cli\Cli;
-use Servebolt\Optimizer\PluginActiveStateHandling\PluginActiveStateHandling;
-
+use function Servebolt\Optimizer\Helpers\featureIsActive;
 use function Servebolt\Optimizer\Helpers\featureIsAvailable;
 use function Servebolt\Optimizer\Helpers\isCli;
+use function Servebolt\Optimizer\Helpers\isFrontEnd;
 use function Servebolt\Optimizer\Helpers\isHostedAtServebolt;
 use function Servebolt\Optimizer\Helpers\isLogin;
 use function Servebolt\Optimizer\Helpers\isTesting;
-use function Servebolt\Optimizer\Helpers\isFrontEnd;
-use function Servebolt\Optimizer\Helpers\featureIsActive;
 
 /**
  * Class ServeboltOptimizer
